+++ src/utilities.ts
@@ -1,4 +1,5 @@
 export { clockToMs } from './clockToMs.js';
+export { msToClock } from './msToClock.js';
 export { msToParts } from './msToParts.js';
 export { parseDuration } from './parseDuration.js';
 export { partsToMs } from './partsToMs.js';